Move rocket show to components module
