+++ apps/fly-my-rockets/src/app/rockets/components/rocket-show/rocket-show.component.scss
@@ -1,9 +1,3 @@
-.fab-button {
-  position: fixed;
-  right: 24px;
-  bottom: 24px;
-}
-
 .flex-header {
   display: flex;
   flex-direction: row;
