+++ apps/fly-my-rockets/src/app/rockets/rockets-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { RocketListComponent } from './components/rocket-list/rocket-list.component';
-import { RocketShowComponent } from './rocket-show/rocket-show.component';
+import { RocketShowComponent } from './components/rocket-show/rocket-show.component';
 
 const routes: Routes = [
   {
