+++ apps/fly-my-rockets/src/app/rockets/rockets.module.ts
@@ -4,17 +4,12 @@ import { CommonModule } from '@angular/common';
 import { RocketsRoutingModule } from './rockets-routing.module';
 import { SharedModule } from '../shared/shared.module';
 import { RocketDialogComponent } from './dialogs/rocket-dialog/rocket-dialog.component';
-import { RocketShowComponent } from './rocket-show/rocket-show.component';
 import { FlightDialogComponent } from './dialogs/flight-dialog/flight-dialog.component';
 import { PipesModule } from './pipes/pipes.module';
 import { ComponentsModule } from './components/components.module';
 
 @NgModule({
-  declarations: [
-    RocketDialogComponent,
-    RocketShowComponent,
-    FlightDialogComponent
-  ],
+  declarations: [RocketDialogComponent, FlightDialogComponent],
   imports: [
     CommonModule,
     RocketsRoutingModule,
